Fixed an error with the email. Previously, if you were to create a settings.ini file and then try to send an email straight after, there would be an authentication error. Now whenever the email to students button is clicked, it checks again for the existence of the settings.ini file. It then uses the settings from that file and fills in the user textfield. This was done with the use of a check settings file method

+++ praCourseWork2/src/praCourseWork2/EmailSettingsFrame.java
@@ -127,7 +127,7 @@ public class EmailSettingsFrame extends JFrame {
 		// Spinner Model
 		Integer value = new Integer(0);
 		Integer min = new Integer(0);
-		Integer max = new Integer(100);
+		Integer max = new Integer(1000);
 		Integer step = new Integer(1);
 		SpinnerModel model = new SpinnerNumberModel(value, min, max, step);
 	    portSpinner = new JSpinner(model);
